ci: add ruff lint job and fix the findings it surfaces
Adds ruff to the dev group and a Lint (ruff) job gated behind the secret
scan, alongside the test job.

Rule selection is scoped to correctness and real code smells rather than
ruff's defaults. The default set flags 98 issues here, but 74 of those are
cosmetic modernisation (UP045 `Optional[X]` -> `X | None` alone accounts
for 66 across 17 files). Auto-fixing all of that would bury a CI change
under ~90 unrelated edits, so UP and I stay out of the gate and belong in
their own reformat PR.

E702 is ignored deliberately: `x = y; changed = True` is consistent style
across 16 sites here, and splitting those lines is `ruff format`'s call.

The 9 real findings are fixed:

- llm_enrich: subprocess.run now passes check=False explicitly (returncode
  was already inspected), and the gcloud-missing RuntimeError raises
  `from None` so the traceback stops blaming FileNotFoundError.
- sec_bulk: dropped duplicate TIBURON / SAN ANSELMO coordinate entries —
  exact repeats of the same values, so no behaviour change; folded a
  DictReader loop into rows.extend; collapsed a two-call startswith into a
  tuple; removed a redundant int() around round().
- wikipedia: removed a redundant int() around round().
- website_enrich: dropped an unused asdict import; the blind except in
  _html_to_text keeps its behaviour with a noqa explaining that arbitrary
  scraped HTML must degrade to "" rather than crash the run.
- tests: removed two unused pytest imports.

89 tests still pass.

+++ scraper/wikipedia.py
@@ -330,7 +330,7 @@ def _parse_aum_string(raw: str) -> Optional[int]:
     elif n < 1000:
         # $-marked small number with no suffix — assume billions in VC context.
         n *= 1_000_000_000
-    return int(round(n))
+    return round(n)
 
 
 def info_from_infobox(title: str, infobox: dict[str, str]) -> WikiInfo:
